Moved api calls into services
